requirements.txt
# Requirements for Enhanced ELIZA Chatbot

# Core dependencies
re>=2.2.1  # For regular expression operations

# AI and Language Model dependencies
langchain>=0.1.0
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=2.0.0
ollama-python>=0.1.0
mcp-python>=0.1.0  # Model Context Protocol
requests>=2.28.0
python-dotenv>=0.19.0
typing-extensions>=4.5.0

# Development dependencies
black>=22.3.0  # Code formatting
isort>=5.10.1  # Import sorting
pylint>=2.17.0  # Code analysis
